Refactor CLI and consolidate server startup logic
- Merge `cli.ts` into `index.ts`
- Add environment variable fallback for token and user
- Update package version to 2.0.0
- Simplify CLI command structure by removing `start` subcommand
- Update README with new CLI usage and configuration options
