+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "pushover-mcp",
-  "version": "1.0.0",
+  "version": "2.0.0",
   "description": "MCP for Pushover.net notifications",
   "author": "Ashik Nesin <mail@ashiknesin.com>",
   "license": "MIT",
@@ -12,7 +12,7 @@
   "type": "module",
   "main": "dist/index.js",
   "bin": {
-    "pushover-mcp": "dist/cli.js"
+    "pushover-mcp": "dist/index.js"
   },
   "files": [
     "dist",
@@ -20,8 +20,8 @@
   ],
   "scripts": {
     "build": "tsc",
-    "start": "node dist/cli.js",
-    "dev": "tsx src/cli.ts",
+    "start": "node dist/index.js",
+    "dev": "tsx src/index.ts",
     "test": "vitest",
     "prepublishOnly": "pnpm build"
   },
