+++ src/index.ts
@@ -1,6 +1,7 @@
 import { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
 import { StdioServerTransport } from "@modelcontextprotocol/sdk/server/stdio.js";
 import { z } from "zod";
+import { Command } from "commander";
 
 const PushoverConfigSchema = z.object({
   token: z.string().min(1),
@@ -28,7 +29,7 @@ export class PushoverMCP {
   constructor() {
     this.server = new McpServer({
       name: "pushover",
-      version: "1.0.0",
+      version: "2.0.0",
       description: "MCP for sending Pushover notifications",
     });
   }
@@ -99,4 +100,68 @@ export class PushoverMCP {
       await this.transport.close();
     }
   }
+}
+
+// Set up Commander CLI
+async function main() {
+  const program = new Command();
+  
+  program
+    .name('pushover-mcp')
+    .description('MCP for sending Pushover notifications')
+    .version('2.0.0')
+    .requiredOption('--token <token>', 'Pushover application token')
+    .requiredOption('--user <user>', 'Pushover user key')
+    .action(async (options) => {
+      const { token, user } = options;
+      let mcp: PushoverMCP | undefined;
+
+      try {
+        console.log('Starting Pushover MCP server...');
+        mcp = new PushoverMCP();
+        await mcp.init({
+          token,
+          user,
+        });
+        
+        console.log('Pushover MCP server started successfully');
+        console.log('Available tools:');
+        console.log('  - send: Send a notification via Pushover');
+        console.log('\nServer is ready to accept commands...');
+
+        // Handle process signals
+        const cleanup = async () => {
+          if (mcp) {
+            console.log('\nShutting down MCP server...');
+            await mcp.close();
+            process.exit(0);
+          }
+        };
+
+        process.on('SIGINT', cleanup);
+        process.on('SIGTERM', cleanup);
+        
+        // Keep the process running
+        await new Promise(() => {});
+      } catch (error) {
+        console.error('Error starting Pushover MCP server:', error);
+        process.exit(1);
+      }
+    });
+
+  // Add support for environment variables as a fallback
+  if (process.env.PUSHOVER_TOKEN && process.env.PUSHOVER_USER && process.argv.length <= 2) {
+    process.argv.push('--token', process.env.PUSHOVER_TOKEN);
+    process.argv.push('--user', process.env.PUSHOVER_USER);
+  }
+
+  await program.parseAsync(process.argv);
+}
+
+// Only run the main function if this is the main module
+if (import.meta.url === `file://${process.argv[1]}`) {
+  main().catch(error => {
+    console.error('Unhandled error:', error);
+    process.exit(1);
+  });
 }
